+++ CHANGELOG.md
@@ -5,7 +5,7 @@ FEATURES:
 * `vcd_vapp_vm` - Ability to add metadata to a VM. For previous behaviour please see `BACKWARDS INCOMPATIBILITIES` [GH-158]
 * `vcd_vapp_vm` - Ability to enable hardware assisted CPU virtualization for VM. It allows hypervisor nesting. [GH-219]
 * **New Resource:** external network - `vcd_external_network` - [GH-230]
-* **New Resource:** VDC resource `vcd_vdc` - [GH-234]
+* **New Resource:** VDC resource `vcd_org_vdc` - [GH-234]
 * resource/vcd_vapp_vm: Add `network` argument for multiple NIC support and more flexible configuration [GH-233]
 * resource/vcd_vapp_vm: Add `mac` argument to store MAC address in state file [GH-233]